backend/student/requirements.txt
# File: requirements.txt (add these for training/evaluation)
# Core Framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
pydantic[email]>=2.5.0

# ML / Recommendations
numpy>=1.26.4
sentence-transformers>=2.2.2
faiss-cpu>=1.8.0
scipy>=1.11.0
scikit-learn>=1.3.0
openai>=1.0.0
google-generativeai>=0.3.0

# Hyperparameter Optimization
optuna>=3.4.0

# Database
beanie>=1.23.1
motor>=3.3.2
pymongo>=4.6.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
argon2-cffi>=23.1.0
python-multipart>=0.0.6
authlib>=1.3.0
itsdangerous>=2.1.2

# Google/Firebase Authentication
google-auth>=2.25.2
google-auth-oauthlib>=1.2.0
firebase-admin>=6.2.0

# 2FA (TOTP)
pyotp>=2.9.0
qrcode[pil]>=7.4.2

# Email
aiosmtplib>=3.0.1
email-validator>=2.1.0

# SMS (Twilio - optional)
twilio>=8.10.0

# Environment & Config
python-dotenv>=1.0.0

# HTTP Client
httpx>=0.25.1

# Redis
redis>=5.0.1

# Date/Time utilities
python-dateutil>=2.8.2

# PDF Generation
fpdf2>=2.7.6

# Pillow (for QR code image generation)
Pillow>=10.1.0

# Visualization (for evaluation)
matplotlib>=3.8.0
seaborn>=0.13.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Progress bars
tqdm>=4.66.0